[BTAT-2544] Refactored connector and view for CustomerInfo to use CustomerDetailsModel, as this is what is returned from vat-subscription
