+++ test/connectors/httpParsers/CustomerDetailsHttpParserSpec.scala
@@ -16,7 +16,7 @@
 
 package connectors.httpParsers
 
-import assets.CustomerInformationTestConstants._
+import assets.CustomerDetailsTestConstants._
 import connectors.httpParsers.CustomerDetailsHttpParser.CustomerDetailsReads
 import models.core.ErrorModel
 import play.api.http.Status
@@ -26,17 +26,16 @@ import utils.TestUtil
 
 class CustomerDetailsHttpParserSpec extends TestUtil {
 
-  val successBadJson = Some(Json.parse("{}"))
+  val successBadJson = Some(Json.obj("firstName" -> 1))
   val errorModel = ErrorModel(Status.BAD_REQUEST, "Error Message")
 
   "The CustomerDetailsHttpParser" when {
 
     "the http response status is OK with valid Json" should {
 
-      "return a CustomerInformationModel" in {
-        CustomerDetailsReads.read("", "", HttpResponse(Status.OK, Some(mtdfbMandatedIndividualJson))) shouldBe Right(mtdfbMandatedIndividual)
+      "return a CustomerDetailsModel" in {
+        CustomerDetailsReads.read("", "", HttpResponse(Status.OK, Some(individualJson))) shouldBe Right(individual)
       }
-
     }
 
     "the http response status is OK with invalid Json" should {
@@ -45,7 +44,6 @@ class CustomerDetailsHttpParserSpec extends TestUtil {
         CustomerDetailsReads.read("", "", HttpResponse(Status.OK, successBadJson)) shouldBe
           Left(ErrorModel(Status.INTERNAL_SERVER_ERROR,"Invalid Json"))
       }
-
     }
 
     "the http response status is BAD_REQUEST" should {
@@ -54,7 +52,6 @@ class CustomerDetailsHttpParserSpec extends TestUtil {
         CustomerDetailsReads.read("", "", HttpResponse(Status.BAD_REQUEST, None)) shouldBe
           Left(ErrorModel(Status.BAD_REQUEST,"Downstream error returned when retrieving CustomerDetails"))
       }
-
     }
 
     "the http response status unexpected" should {
@@ -63,10 +60,6 @@ class CustomerDetailsHttpParserSpec extends TestUtil {
         CustomerDetailsReads.read("", "", HttpResponse(Status.SEE_OTHER, None)) shouldBe
           Left(ErrorModel(Status.SEE_OTHER,"Downstream error returned when retrieving CustomerDetails"))
       }
-
     }
   }
-
-
-
 }
