--- test/models/customerInfo/CustomerInformationModelSpec.scala
@@ -1,53 +1,0 @@
-/*
- * Copyright 2018 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package models.customerInfo
-
-import assets.BaseTestConstants._
-import assets.CustomerInformationTestConstants._
-import models.customerInfo.CustomerInformationModel._
-import play.api.libs.json.Json
-import uk.gov.hmrc.play.test.UnitSpec
-
-
-class CustomerInformationModelSpec extends UnitSpec {
-
-  "CustomerInformationModel" should {
-
-    "Deserialize from JSON" when {
-
-      "all optional fields are populated" in {
-        customerInformationJsonMax.as[CustomerInformationModel] shouldBe customerInformationMax
-      }
-
-      "no optional fields are returned" in {
-        customerInformationJsonMin.as[CustomerInformationModel] shouldBe customerInformationMin
-      }
-    }
-
-    "Serialize to JSON" when {
-
-      "all optional fields are populated" in {
-        Json.toJson(customerInformationMax) shouldBe customerInformationJsonMax
-      }
-
-      "no optional fields are returned" in {
-        Json.toJson(customerInformationMin) shouldBe customerInformationJsonMin
-      }
-    }
-  }
-
-}
